refactor: move out reaction list to own file
